Refactor AudioNowPlayingActivity to fragment
